+++ src/core/config.h
@@ -129,7 +129,7 @@ struct RadioConfig {
     // keep their old behavior unless a user opts in.
     uint8_t jitterMs = 0;      // 0..20: random ms between deauth/disassoc (anti-WIDS)
     uint8_t cooldownMs = 0;    // 0..30s: per-AP cooldown after kick (PORKCHOP method)
-    int8_t  scoreThr = 0;      // -100..200: min score to attack in PORKCHOP method (0 = score all)
+    int16_t scoreThr = 0;      // -100..200: min score to attack in PORKCHOP method (0 = score all)
     uint16_t dwellMinMs = 120; // 50..600: minimum channel dwell (for PASSIVE-like adaptive hop)
 };
